requirements.txt
pika
Flask
flask-cors
requests
kafka-python-ng
fluent-logger